Add "out" statements
Add support for "out <expr>" statements. Previously expressions acted as
program units and evaluation result would be printed in the out output. This
commit introduces notion of "statement" which is the top level unit inside
the program. For there are only "out" statements, which effectively do the
same as the previous behaviour - print result of expression which is passed
as an argument.

Test cases are updated to reflect requirements to new syntax.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>