Enable verbose error reporting in bison
By default if something goes wrong, bison only print "Syntax error"
message and location of error (if that was enabled), for example see GNU
ld linker when it finds some error in linker script files. With
YYERROR_VERBOSE bison would at least print what it got and what it
expected, although that is still pretty cryptic for the end user.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>