Disable bogus build/include warning in cpplint
Warning build/include demands that header files included with quotes
should be referenced by the directory name. This error doesn't make a
lot of sense in case when header is already in the same directory. And
prefixing header with "./" just makes things worse, because now cpplint
complains that it is "./" instead of "interpreter".