Supports comments in program
This change adds support for typical comments in code: those that start
with `#' and end at the end of the line. To support this it is required
to change lexer to skip such lines, and to parser to ignore empty lines.
Also main.cc should have be updated to ensure that "Statement" instance
returned from the parser is always null, for comment lines.

I also considered an implementation where bison parser whole program at
once, instead of line by line, however that made implementation a bit
more complex both in the parser and in the invoking code, which seemed
unjustified to me, because grammar doesn't need this - it is strictly
one statement per line, one line per statement.