Fix line and column number handling
It appears that structure that tracks column and linen umbers is cached
between parser invocations, as a result in some cases column number
wasn't properly reset to 0 on a new line. And in some cases it was just
a bogus number that was clearly too big to be a column number. This can
be fixed by setting column number explicitly to 0. There is no public
yyset_column function like yyset_lineno, so instead buffer structure
should be changed directly.

Also while looking at the internals of parser it appeared that by
default while columns are started at 0, lines start at 1. This breaks my
previous assumption that both should be either 0-based or 1-based, but
not inconsistently different. To avoid problems with generated code it
makes sense to change my code also to start lines at 1.