Fix lint issues in header files
cpplint parses only files that were explicitly specified, so header
files of this project were left unchecked. This commit adds header files
to cpplint invocation and fixes issues reported by it. This patch also
renames data-members to follow Google Coding Style - name should end
with underscore. This is an error that is not reported by cpplint, I
guess because it is a simple regexp script, rather than proper semantic
analyzer like clang-tidy (which blows it in other aspects, OTOH).