Handle exceptions from the statement execution
If there is an exception during statement execution (for example
reference to unknown identifier), then message from this exception will
be printed together with the line number. At the moment of execution we
don't have an information about columns for a given symbol, so it isn't
possible to print column info for error source.