Add support for variables
This change adds supprot for variables. Values are assigned via

    var <identifier> = <expr>

statements. Variables can be used as is in expressions. If there is an
attempt to dereference unknown variable, then there is an error message
and execution is aborted. Variables can be reassigned and assignments
can be recursive: expression can reference current value of this
variable.