Update test cases for missign identifier case
Syntax errors don't halt execution - it continues to move on so that it
would find as many syntax errors is possible. But if there is an
execution error, then execution stops immediately, to avoid any invalid
results (for example if variable is reassigned, further references to
this variable would be valid and will be executed, but will use wrong
value). Some of the test cases that check for undefined variable
references should be updated therefore.

There is actually a problem here, because lines are parsed by one, so in
the previous example of variable assignment, if there is a syntax error
in the line, then the line will be skipped, but execution will continue.
So in the current arrangment, if there was a syntax line, then
interpereter should move on with syntax parsing of program, but shall
not execute the statements - this will be fixed in the further patch.