Don't execute statements after a syntax error
This commit makes the change described in the prevous commit - now, if
there is a syntax error interpreter will continue to parse lines, but
will not try to execute them, thus avoiding the bogus results.

It also introduces similar behaviour for cases of execution errors -
interpreter will continue to parse lines after them.