Generate dependency files
Header files were not listed as dependencies for source files, since
maintaining proper list of dependencies would be difficult. However
that was causing a problem, because now if header file was changed, make
wouldn't rebuild dependant source files. One simple way to fix that
would be to list as all source files being dependent on all header
files, but that would trigger full rebuild in 100% of header file
changes, which is not always needed. Instead we can use gcc's option
-MMD to generate files with list of dependant header files, and then
include those files in Makefile, thus making sure that dependencies are
known to make.