Change extension of generated files to .cc
lexer.l and parser.y used to generate .c files, but that was causing
complications in the Makefile, which had to handle both .c and .cc.
Technically those files are still just a C, but they were already
compiled with -std=c++11 anyway, so there are no changes in the way how
files are treated by the compiler.