Add support for range values
This comment adds supports for range values and expressions. Value is
still a single class that can contain either scalar integer, scalar
float or integer range inside, depending on `type_` field. At this point
there are good reasons to replace this with several classes with virtual
function-members like `asString()`, however I've found that this
implementation is quite more complex, since now Value cannot be passed
by value, but should be passed by the pointer, so there is a need for
memory management, either via new/delete or via std::shared_ptr. I've
tried std::unique_ptr, but that worked quite badly because of the
variables stored in the Context class, since they need to be copied.
Size of the current Value class didn't increased in most cases, because
`double` is twice the `int` on most platforms. If we'd need to support
some another value type, than very likely it would be needed to redesign
this Value class to become hierarchy with virtual function-members.

This patch doesn't implement proper handling of cases where user tries
to do some operation on scalar and range values (that should throw an
error) or on two range operations (this is undefined in grammar,
probably should trigger an error too).