Add support for sequence values
This commits VectorValue class that represents vectors. This is
different from range values: range values are a range of integer numbers
from begin to end with a step of 1. Vector values on the other hand
represent an arbitrary sequence of numbers. As a result, while range cam
be implemeneted very efficiently (just store current value and end
value), vector has to store all of it's values. Current implement is
simple and just stores everything in the std::vector, also all ValuePtr
instances should exist at the moment of VectorValue creation.

It wouldn't hurt to split original Value class into at least ScalarValue
and RangeValue (it is also possible to come up with a smart way to merge
Range and Vector, by adding some generator function to Vector, that
potentially might help with support evaluation over super large memory
sets, but that is for the future).