Change access level for Value's isScalar and isScalarFloat
It appears that I can't call protected methods from inheriting classes
if that is done on another instance of the base class rather then on
`this` - or at least for some reason GCC throws a syntax error in such
scenario. As a result some protected function-members should be made
public to enable decorators.