Use shared value vector in VectorValue
Initial implementation of VectorValue was using a unique copy of
std::vector to store values. That was causing an performance issues with
iterating, because each iteration step was doing a complete copy of
vector, and that became quite a prominent drag on performance with large
N. This commit addresses this problems by using the same shared instance
of vector across all VectorValues which point to any value in this
vector.