Concurrent reduce execution
Similar to the previous commit, this implements the true concurrent
execution of reduce expressions. Similarly it also doesn't implement any
optimization for small input set, where overhead of multithreading will
be bigger than any advantage, although it is not hard to implement such
optimiztion if needed, but it is much harder to identify at which value
to put the red line, as actual values might depend on particular CPU and
OS in use.