Remove obsolete AsyncValue class
This class was born from the improper understanding of how map/reduce
should be calculated in parallel and it doesn't really make any sense
any more. It's purpose was to allow evaluate of expressions in
background, while other expressions are evaluated, but in most cases
result is needed immediately and is calculated quickly, so there is no
place for backgroundness. And in case of map/reduce where parallel
execution makes sense this achieved via other means.