Throw error if lambda returns a non scalar value
OTOH it would be fun to have vectors of vectors and use nested maps and
reduces. I've tried this and it mostly worked, but get broken in case of
nested reduced calls like

    var a = map( {1, 2}, z -> {z, z * 2} )
    var b = reduce( a, { 0 1 },
	z x -> reduce(z, 0, q w -> q + w) * reduce(x, 0, r t -> r + t))

But if nested reduce is done first in a separate step, current code
works quite fine though. I haven't tried to understand how exactly that
can be fixed, so for now lets assume that only single-dimension vectors
are supported.